separate add to list to new component
