+++ src/ReadonlyRecipe.js
@@ -1,40 +1,18 @@
-import React, { useState, useEffect, useContext } from 'react'
-import { Button, Input, Badge, InputGroup, InputGroupAddon, InputGroupText, Alert } from 'reactstrap'
+import React, { useState, useContext } from 'react'
+import { Badge } from 'reactstrap'
 import './Styles.css'
 import iconServings from './icons/icons8-restaurant-24.png'
 import iconEdit from './icons/icons8-edit-24.png'
 import iconDelete from './icons/icons8-trash-can-24.png'
+import { AddRecipeToListForm } from './AddRecipeToListForm'
 import { StarRating } from './StarRating'
 import { Link } from 'react-router-dom'
 import { UserContext } from './UserContext'
 import { api } from './services/api'
 
 export function ReadonlyRecipe (props) {
-  const [lists, setLists] = useState([])
-  const [selectedList, setSelectedList] = useState('')
-  const [listServings, setListServings] = useState(props.recipe.servings)
-  const [newListName, setNewListName] = useState('')
-  const [newListId, setNewListId] = useState()
-  const [addedToListAlertVisible, setAddedToListAlertVisible] = useState(false)
   const [deletedRecipe, setDeletedRecipe] = useState(false)
-
-  const addedToListAlertOnDismiss = () => setAddedToListAlertVisible(false)
-
   const idToken = useContext(UserContext)
-
-  useEffect(() => {
-    api.getLists(idToken)
-      .then((response) => {
-        const dbLists = response.data
-        let listNames = []
-        listNames = dbLists.map((list) => {
-          return { id: list._id, title: list.title, recipesUrl: list.recipes.href }
-        })
-        listNames.push({ id: 'newlist', title: 'New list' })
-        setLists(listNames)
-        setSelectedList(listNames[0].id)
-      })
-  }, [idToken])
 
   const recipe = props.recipe
   let listTags
@@ -62,34 +40,6 @@ export function ReadonlyRecipe (props) {
     event.target.style.display = 'none'
   }
 
-  const handleAddToList = (event) => {
-    event.preventDefault()
-    addedToListAlertOnDismiss()
-
-    const listObject = {
-      recipeId: recipe._id,
-      recipeServings: listServings
-    }
-
-    if (selectedList !== 'newlist') {
-      return api.addRecipeToList(idToken, selectedList, listObject)
-        .then(() => {
-          setAddedToListAlertVisible(true)
-        })
-    } else {
-      const newList = { title: newListName }
-      return api.createList(idToken, newList)
-        .then((response) => {
-          const dbList = response.data
-          setNewListId(dbList._id)
-          return api.addRecipeToList(idToken, dbList._id, listObject)
-            .then(() => {
-              setAddedToListAlertVisible(true)
-            })
-        })
-    }
-  }
-
   const handleDeleteRecipe = () => {
     return api.deleteRecipe(idToken, recipe._id)
       .then(() => {
@@ -97,44 +47,12 @@ export function ReadonlyRecipe (props) {
       })
   }
 
-  let listOptions = []
-  listOptions = lists.map((list) => {
-    return <option key={list.id} value={list.id}>{list.title}</option>
-  })
-
-  const newListInput = (
-    <Input type='text' name='listName' id='listNameText' value={newListName} onChange={(event) => setNewListName(event.target.value)} placeholder='New list name' />
-  )
-
-  const listServingsInput = (
-    <>
-      <InputGroupAddon addonType='prepend'>
-        <InputGroupText>Servings:</InputGroupText>
-      </InputGroupAddon>
-      <Input type='text' name='listServings' id='servingsText' value={listServings} onChange={(event) => setListServings(event.target.value)} placeholder='Servings to add to list' />
-    </>
-  )
-
-  const listUrl = selectedList !== 'newlist' ? `/lists/${selectedList}` : `/lists/${newListId}`
-  let selectedListName
-  lists.forEach((list) => {
-    if (list.id === selectedList) {
-      selectedListName = list.title
-    }
-  })
-  const listName = selectedList !== 'newlist' ? selectedListName : newListName
-  const addedToListAlert = (
-    <Alert color='info' isOpen={addedToListAlertVisible} toggle={addedToListAlertOnDismiss}>
-      Recipe added to list <Link to={listUrl}>{listName}</Link>
-    </Alert>
-  )
-
   return (
     <div>
       <div className='recipeHeaderContainer'>
         {recipe.image
           ? <img src={recipe.image} alt={recipe.title} onError={handleImageError} className='recipeHeaderImage' />
-          : null }
+          : null}
         <div>
           <h2>{recipe.title}
             <span className='actionsMenu'>
@@ -158,23 +76,7 @@ export function ReadonlyRecipe (props) {
             : null}
           {listTags}
           <hr />
-          <h6>Add to shopping list</h6>
-          <InputGroup>
-            <InputGroupAddon addonType='prepend'>
-              <InputGroupText>List:</InputGroupText>
-            </InputGroupAddon>
-            <Input type='select' name='listsSelect' id='listsSelect' value={selectedList} onChange={(event) => setSelectedList(event.target.value)}>
-              {listOptions}
-            </Input>
-            {selectedList === 'newlist'
-              ? newListInput
-              : null}
-            {recipe.servings
-              ? listServingsInput
-              : null}
-            <InputGroupAddon addonType='append'><Button type='submit' id='addRecipeToListButton' aria-label='add recipe to list' onClick={handleAddToList}>Add</Button></InputGroupAddon>
-          </InputGroup>
-          {addedToListAlert}
+          <AddRecipeToListForm recipeId={recipe._id} recipeServings={recipe.servings} />
         </div>
       </div>
       <br />
